# demo/src/main/resources/swagger.yml
{
  "swagger": "2.0",
  "info": {
    "description": "YourBank API",
    "version": "1.0.0",
    "title": "YourBank API"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "your-bank-controller",
      "description": "Your Bank Controller"
    }
  ],
  "paths": {
    "/digital/yourbank/accounts/{accountId}/transactions": {
      "get": {
        "tags": [
          "your-bank-controller"
        ],
        "summary": "Get all linked accounts by userId",
        "operationId": "getTransactionsUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "accountId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Transaction"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/digital/yourbank/customer/login": {
      "post": {
        "tags": [
          "your-bank-controller"
        ],
        "summary": "login",
        "operationId": "loginUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "loginCredential",
            "description": "loginCredential",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful login",
            "schema": {
              "$ref": "#/definitions/CustomerProfile"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Customer not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/digital/yourbank/customer/register": {
      "post": {
        "tags": [
          "your-bank-controller"
        ],
        "summary": "register user",
        "operationId": "registerUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "request",
            "description": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user registered successfully",
            "schema": {
              "$ref": "#/definitions/CustomerProfile"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/digital/yourbank/customer/{userId}/account": {
      "post": {
        "tags": [
          "your-bank-controller"
        ],
        "summary": "open account",
        "operationId": "openAccountUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "request",
            "description": "openAccountRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OpenAccountRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "account opened successfully",
            "schema": {
              "$ref": "#/definitions/AccountSummary"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "your-bank-controller"
        ],
        "summary": "link account to the user",
        "operationId": "linkAccountUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "employeeDetails",
            "description": "linkAccountRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LinkAccountRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AccountSummary"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/digital/yourbank/customer/{userId}/accounts": {
      "get": {
        "tags": [
          "your-bank-controller"
        ],
        "summary": "Get all linked accounts by userId",
        "operationId": "getAllLinkedAccountsUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AccountSummary"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/digital/yourbank/customer/{userId}/fund": {
      "post": {
        "tags": [
          "your-bank-controller"
        ],
        "summary": "transferFund",
        "operationId": "transferFundUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "request",
            "description": "transferFundRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TransferFundRequestVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "fund transferred successfully",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Transaction"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Account": {
      "type": "object",
      "properties": {
        "accountId": {
          "type": "string"
        },
        "accountToken": {
          "type": "string"
        },
        "balance": {
          "type": "number",
          "format": "double"
        },
        "displayAccountNumber": {
          "type": "string"
        },
        "limit": {
          "type": "number",
          "format": "double"
        }
      },
      "title": "Account"
    },
    "AccountSummary": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/Account"
        },
        "accountDescription": {
          "type": "string"
        },
        "associationType": {
          "type": "string"
        },
        "ownerDetail": {
          "$ref": "#/definitions/CustomerProfile"
        }
      },
      "title": "AccountSummary"
    },
    "CustomerProfile": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "title": "CustomerProfile"
    },
    "LinkAccountRequest": {
      "type": "object",
      "properties": {
        "accountToken": {
          "type": "string"
        }
      },
      "title": "LinkAccountRequest"
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "title": "LoginRequest"
    },
    "OpenAccountRequest": {
      "type": "object",
      "properties": {
        "accountDescription": {
          "type": "string"
        },
        "limit": {
          "type": "number",
          "format": "double"
        }
      },
      "title": "OpenAccountRequest"
    },
    "RegisterUserRequest": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "title": "RegisterUserRequest"
    },
    "Transaction": {
      "type": "object",
      "properties": {
        "contributor": {
          "$ref": "#/definitions/CustomerProfile"
        },
        "transactionAmount": {
          "type": "number",
          "format": "double"
        },
        "transactionDescription": {
          "type": "string"
        },
        "transactionId": {
          "type": "string"
        },
        "transactionType": {
          "type": "string"
        }
      },
      "title": "Transaction"
    },
    "TransferFundRequestVO": {
      "type": "object",
      "properties": {
        "accountId": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "double"
        }
      },
      "title": "TransferFundRequestVO"
    }
  }
}